+++ robotframework-ls/tests/robotframework_ls_tests/_resources/case_params_on_lib/LibWithParams3.py
@@ -1,0 +1,6 @@
+class LibWithParams3:
+    def __init__(self, resolved_name):
+        assert resolved_name == "RESOLVED"
+
+    def check(self):
+        return 1