+++ robotframework-ls/tests/robotframework_ls_tests/test_code_analysis.py
@@ -363,6 +363,54 @@ def test_code_analysis_lib_with_params(
     _collect_errors(workspace, doc, data_regression, basename="no_error", config=config)
 
 
+def test_code_analysis_lib_with_params_2(
+    workspace, libspec_manager, cases, data_regression
+):
+    from robotframework_ls.robot_config import RobotConfig
+
+    workspace.set_root("case_params_on_lib", libspec_manager=libspec_manager)
+
+    caseroot = cases.get_path("case_params_on_lib")
+    config = RobotConfig()
+    config.update(
+        {
+            "robot": {
+                "pythonpath": [caseroot],
+                "libraries": {"libdoc": {"needsArgs": ["*"]}},
+            }
+        }
+    )
+    libspec_manager.config = config
+
+    doc = workspace.get_doc("case_params_on_lib3.robot")
+
+    _collect_errors(workspace, doc, data_regression, basename="no_error", config=config)
+
+
+def test_code_analysis_lib_with_params_3(
+    workspace, libspec_manager, cases, data_regression
+):
+    from robotframework_ls.robot_config import RobotConfig
+
+    workspace.set_root("case_params_on_lib", libspec_manager=libspec_manager)
+
+    caseroot = cases.get_path("case_params_on_lib")
+    config = RobotConfig()
+    config.update(
+        {
+            "robot": {
+                "pythonpath": [caseroot],
+                "libraries": {"libdoc": {"needsArgs": ["*"]}},
+            }
+        }
+    )
+    libspec_manager.config = config
+
+    doc = workspace.get_doc("case_params_on_lib4.robot")
+
+    _collect_errors(workspace, doc, data_regression, basename="no_error", config=config)
+
+
 def test_code_analysis_same_lib_multiple_with_alias(
     workspace, libspec_manager, data_regression
 ):